+++ src/lib/__tests__/utils-normalizeConfig-test.js
@@ -11,14 +11,14 @@
 
 jest.autoMockOff();
 
-describe('utils-normalizeConfig', function() {
-  var path;
-  var root;
-  var utils;
-  var expectedPathFooBar;
-  var expectedPathFooQux;
-  var expectedPathAbs;
-  var expectedPathAbsAnother;
+describe('utils-normalizeConfig', () => {
+  let path;
+  let root;
+  let utils;
+  let expectedPathFooBar;
+  let expectedPathFooQux;
+  let expectedPathAbs;
+  let expectedPathAbsAnother;
 
   // Windows uses backslashes for path separators, which need to be escaped in
   // regular expressions. This little helper function helps us generate the
@@ -30,7 +30,7 @@ describe('utils-normalizeConfig', function() {
     );
   }
 
-  beforeEach(function() {
+  beforeEach(() => {
     path = require('path');
     root = path.resolve('/');
     expectedPathFooBar = path.join(root, 'root', 'path', 'foo', 'bar', 'baz');
@@ -40,9 +40,9 @@ describe('utils-normalizeConfig', function() {
     utils = require('../utils');
   });
 
-  it('throws when an invalid config option is passed in', function() {
+  it('throws when an invalid config option is passed in', () => {
 
-    expect(function() {
+    expect(() => {
       utils.normalizeConfig({
         rootDir: '/root/path/foo',
         thisIsAnInvalidConfigKey: 'with a value even!',
@@ -51,17 +51,17 @@ describe('utils-normalizeConfig', function() {
 
   });
 
-  describe('rootDir', function() {
-    it('throws if the config is missing a rootDir property', function() {
-      expect(function() {
+  describe('rootDir', () => {
+    it('throws if the config is missing a rootDir property', () => {
+      expect(() => {
         utils.normalizeConfig({});
       }).toThrow(new Error('No rootDir config value found!'));
     });
   });
 
-  describe('collectCoverageOnlyFrom', function() {
-    it('normalizes all paths relative to rootDir', function() {
-      var config = utils.normalizeConfig({
+  describe('collectCoverageOnlyFrom', () => {
+    it('normalizes all paths relative to rootDir', () => {
+      const config = utils.normalizeConfig({
         rootDir: '/root/path/foo/',
         collectCoverageOnlyFrom: {
           'bar/baz': true,
@@ -69,15 +69,15 @@ describe('utils-normalizeConfig', function() {
         },
       }, '/root/path');
 
-      var expected = {};
+      const expected = {};
       expected[expectedPathFooBar] = true;
       expected[expectedPathFooQux] = true;
 
       expect(config.collectCoverageOnlyFrom).toEqual(expected);
     });
 
-    it('does not change absolute paths', function() {
-      var config = utils.normalizeConfig({
+    it('does not change absolute paths', () => {
+      const config = utils.normalizeConfig({
         rootDir: '/root/path/foo',
         collectCoverageOnlyFrom: {
           '/an/abs/path': true,
@@ -85,31 +85,31 @@ describe('utils-normalizeConfig', function() {
         },
       });
 
-      var expected = {};
+      const expected = {};
       expected[expectedPathAbs] = true;
       expected[expectedPathAbsAnother] = true;
 
       expect(config.collectCoverageOnlyFrom).toEqual(expected);
     });
 
-    it('substitutes <rootDir> tokens', function() {
-      var config = utils.normalizeConfig({
+    it('substitutes <rootDir> tokens', () => {
+      const config = utils.normalizeConfig({
         rootDir: '/root/path/foo',
         collectCoverageOnlyFrom: {
           '<rootDir>/bar/baz': true,
         },
       });
 
-      var expected = {};
+      const expected = {};
       expected[expectedPathFooBar] = true;
 
       expect(config.collectCoverageOnlyFrom).toEqual(expected);
     });
   });
 
-  describe('testPathDirs', function() {
-    it('normalizes all paths relative to rootDir', function() {
-      var config = utils.normalizeConfig({
+  describe('testPathDirs', () => {
+    it('normalizes all paths relative to rootDir', () => {
+      const config = utils.normalizeConfig({
         rootDir: '/root/path/foo',
         testPathDirs: [
           'bar/baz',
@@ -122,8 +122,8 @@ describe('utils-normalizeConfig', function() {
       ]);
     });
 
-    it('does not change absolute paths', function() {
-      var config = utils.normalizeConfig({
+    it('does not change absolute paths', () => {
+      const config = utils.normalizeConfig({
         rootDir: '/root/path/foo',
         testPathDirs: [
           '/an/abs/path',
@@ -136,8 +136,8 @@ describe('utils-normalizeConfig', function() {
       ]);
     });
 
-    it('substitutes <rootDir> tokens', function() {
-      var config = utils.normalizeConfig({
+    it('substitutes <rootDir> tokens', () => {
+      const config = utils.normalizeConfig({
         rootDir: '/root/path/foo',
         testPathDirs: [
           '<rootDir>/bar/baz',
@@ -148,9 +148,9 @@ describe('utils-normalizeConfig', function() {
     });
   });
 
-  describe('scriptPreprocessor', function() {
-    it('normalizes the path according to rootDir', function() {
-      var config = utils.normalizeConfig({
+  describe('scriptPreprocessor', () => {
+    it('normalizes the path according to rootDir', () => {
+      const config = utils.normalizeConfig({
         rootDir: '/root/path/foo',
         scriptPreprocessor: 'bar/baz',
       }, '/root/path');
@@ -158,8 +158,8 @@ describe('utils-normalizeConfig', function() {
       expect(config.scriptPreprocessor).toEqual(expectedPathFooBar);
     });
 
-    it('does not change absolute paths', function() {
-      var config = utils.normalizeConfig({
+    it('does not change absolute paths', () => {
+      const config = utils.normalizeConfig({
         rootDir: '/root/path/foo',
         scriptPreprocessor: '/an/abs/path',
       });
@@ -167,8 +167,8 @@ describe('utils-normalizeConfig', function() {
       expect(config.scriptPreprocessor).toEqual(expectedPathAbs);
     });
 
-    it('substitutes <rootDir> tokens', function() {
-      var config = utils.normalizeConfig({
+    it('substitutes <rootDir> tokens', () => {
+      const config = utils.normalizeConfig({
         rootDir: '/root/path/foo',
         scriptPreprocessor: '<rootDir>/bar/baz',
       });
@@ -177,38 +177,9 @@ describe('utils-normalizeConfig', function() {
     });
   });
 
-  describe('setupEnvScriptFile', function() {
-    it('normalizes the path according to rootDir', function() {
-      var config = utils.normalizeConfig({
-        rootDir: '/root/path/foo',
-        setupEnvScriptFile: 'bar/baz',
-      }, '/root/path');
-
-      expect(config.setupEnvScriptFile).toEqual(expectedPathFooBar);
-    });
-
-    it('does not change absolute paths', function() {
-      var config = utils.normalizeConfig({
-        rootDir: '/root/path/foo',
-        setupEnvScriptFile: '/an/abs/path',
-      });
-
-      expect(config.setupEnvScriptFile).toEqual(expectedPathAbs);
-    });
-
-    it('substitutes <rootDir> tokens', function() {
-      var config = utils.normalizeConfig({
-        rootDir: '/root/path/foo',
-        setupEnvScriptFile: '<rootDir>/bar/baz',
-      });
-
-      expect(config.setupEnvScriptFile).toEqual(expectedPathFooBar);
-    });
-  });
-
-  describe('setupTestFrameworkScriptFile', function() {
-    it('normalizes the path according to rootDir', function() {
-      var config = utils.normalizeConfig({
+  describe('setupTestFrameworkScriptFile', () => {
+    it('normalizes the path according to rootDir', () => {
+      const config = utils.normalizeConfig({
         rootDir: '/root/path/foo',
         setupTestFrameworkScriptFile: 'bar/baz',
       }, '/root/path');
@@ -216,8 +187,8 @@ describe('utils-normalizeConfig', function() {
       expect(config.setupTestFrameworkScriptFile).toEqual(expectedPathFooBar);
     });
 
-    it('does not change absolute paths', function() {
-      var config = utils.normalizeConfig({
+    it('does not change absolute paths', () => {
+      const config = utils.normalizeConfig({
         rootDir: '/root/path/foo',
         setupTestFrameworkScriptFile: '/an/abs/path',
       });
@@ -225,8 +196,8 @@ describe('utils-normalizeConfig', function() {
       expect(config.setupTestFrameworkScriptFile).toEqual(expectedPathAbs);
     });
 
-    it('substitutes <rootDir> tokens', function() {
-      var config = utils.normalizeConfig({
+    it('substitutes <rootDir> tokens', () => {
+      const config = utils.normalizeConfig({
         rootDir: '/root/path/foo',
         setupTestFrameworkScriptFile: '<rootDir>/bar/baz',
       });
@@ -235,11 +206,40 @@ describe('utils-normalizeConfig', function() {
     });
   });
 
-  describe('testPathIgnorePatterns', function() {
-    it('does not normalize paths relative to rootDir', function() {
+  describe('setupTestFrameworkScriptFile', () => {
+    it('normalizes the path according to rootDir', () => {
+      const config = utils.normalizeConfig({
+        rootDir: '/root/path/foo',
+        setupTestFrameworkScriptFile: 'bar/baz',
+      }, '/root/path');
+
+      expect(config.setupTestFrameworkScriptFile).toEqual(expectedPathFooBar);
+    });
+
+    it('does not change absolute paths', () => {
+      const config = utils.normalizeConfig({
+        rootDir: '/root/path/foo',
+        setupTestFrameworkScriptFile: '/an/abs/path',
+      });
+
+      expect(config.setupTestFrameworkScriptFile).toEqual(expectedPathAbs);
+    });
+
+    it('substitutes <rootDir> tokens', () => {
+      const config = utils.normalizeConfig({
+        rootDir: '/root/path/foo',
+        setupTestFrameworkScriptFile: '<rootDir>/bar/baz',
+      });
+
+      expect(config.setupTestFrameworkScriptFile).toEqual(expectedPathFooBar);
+    });
+  });
+
+  describe('testPathIgnorePatterns', () => {
+    it('does not normalize paths relative to rootDir', () => {
       // This is a list of patterns, so we can't assume any of them are
       // directories
-      var config = utils.normalizeConfig({
+      const config = utils.normalizeConfig({
         rootDir: '/root/path/foo',
         testPathIgnorePatterns: [
           'bar/baz',
@@ -253,10 +253,10 @@ describe('utils-normalizeConfig', function() {
       ]);
     });
 
-    it('does not normalize trailing slashes', function() {
+    it('does not normalize trailing slashes', () => {
       // This is a list of patterns, so we can't assume any of them are
       // directories
-      var config = utils.normalizeConfig({
+      const config = utils.normalizeConfig({
         rootDir: '/root/path/foo',
         testPathIgnorePatterns: [
           'bar/baz',
@@ -270,8 +270,8 @@ describe('utils-normalizeConfig', function() {
       ]);
     });
 
-    it('substitutes <rootDir> tokens', function() {
-      var config = utils.normalizeConfig({
+    it('substitutes <rootDir> tokens', () => {
+      const config = utils.normalizeConfig({
         rootDir: '/root/path/foo',
         testPathIgnorePatterns: [
           'hasNoToken',
@@ -286,11 +286,11 @@ describe('utils-normalizeConfig', function() {
     });
   });
 
-  describe('modulePathIgnorePatterns', function() {
-    it('does not normalize paths relative to rootDir', function() {
+  describe('modulePathIgnorePatterns', () => {
+    it('does not normalize paths relative to rootDir', () => {
       // This is a list of patterns, so we can't assume any of them are
       // directories
-      var config = utils.normalizeConfig({
+      const config = utils.normalizeConfig({
         rootDir: '/root/path/foo',
         modulePathIgnorePatterns: [
           'bar/baz',
@@ -304,10 +304,10 @@ describe('utils-normalizeConfig', function() {
       ]);
     });
 
-    it('does not normalize trailing slashes', function() {
+    it('does not normalize trailing slashes', () => {
       // This is a list of patterns, so we can't assume any of them are
       // directories
-      var config = utils.normalizeConfig({
+      const config = utils.normalizeConfig({
         rootDir: '/root/path/foo',
         modulePathIgnorePatterns: [
           'bar/baz',
@@ -321,8 +321,8 @@ describe('utils-normalizeConfig', function() {
       ]);
     });
 
-    it('substitutes <rootDir> tokens', function() {
-      var config = utils.normalizeConfig({
+    it('substitutes <rootDir> tokens', () => {
+      const config = utils.normalizeConfig({
         rootDir: '/root/path/foo',
         modulePathIgnorePatterns: [
           'hasNoToken',