+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 * Fixed unmocking behavior with npm3.
 * Fixed test runtime error reporting and stack traces.
 * Jest sets `process.NODE_ENV` to `test` unless otherwise specified.
+* Added `babel-plugin-jest-unmock`, `babel-jest-preset` and `babel-jest`
+  as packages. `babel-jest` is now being auto-detected.
 
 ## 0.8.2